requirements.txt
blinker==1.9.0
boto3==1.36.4
botocore==1.36.4
click==8.1.8
Flask==3.1.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.1
six==1.17.0
urllib3==1.26.20
uWSGI==2.0.28
Werkzeug==3.1.3
zipp==3.21.0
